added a table like structure and header to the Calendar
